UI/InputViewControls: rename files
