+++ components/ILIAS/UI/resources/js/Input/ViewControl/src/sortation.js
@@ -13,7 +13,7 @@
  * https://github.com/ILIAS-eLearning
  */
 
-export default class ModeFactory {
+export default class Sortation {
   /**
    * @type {JQueryEventDispatcher}
    */
@@ -28,22 +28,28 @@ export default class ModeFactory {
 
   /**
    * @param {HTMLElement} component
-   * @param {string} optValue
+   * @param {string} internalSignal
    * @param {string} containerSubmitSignal
    * @return {void}
    */
-  init(component, optValue, containerSubmitSignal) {
-    component.addEventListener(
-      'click',
-      (event) => {
-        const btn = event.srcElement;
-        btn.parentElement.querySelectorAll('button').forEach(
-          (button) => button.classList.remove('engaged'),
-        );
-        btn.classList.add('engaged');
-        btn.closest('.il-viewcontrol')
-          .querySelector('.il-viewcontrol-value > input')
-          .value = optValue;
+  init(component, internalSignal, containerSubmitSignal) {
+    this.#eventDispatcher.register(
+      component.ownerDocument,
+      internalSignal,
+      (event, signalData) => {
+        let container = event.target.closest('.il-viewcontrol-fieldselection');
+        if (signalData.options.parent_container) {
+          container = component.ownerDocument.querySelector(
+            `#${signalData.options.parent_container
+            } .il-viewcontrol-sortation`,
+          );
+        } else {
+          container = event.target.closest('.il-viewcontrol-sortation');
+        }
+        const inputs = container.querySelectorAll('.il-viewcontrol-value > input');
+        const val = signalData.options.value.split(':');
+        [inputs[0].value, inputs[1].value] = val;
+
         this.#eventDispatcher.dispatch(event.target, containerSubmitSignal);
         return false;
       },
